requirements.txt
Flask==2.2.5
Werkzeug==3.0.1
numpy==1.23.5
requests==2.31.0
gunicorn==21.2.0
scikit-learn==1.2.2
Pillow==9.4.0
#tensorflow-cpu==2.14.0
tensorflow-cpu==2.11.0